// src/shared/types/permissions.ts
export type AppTeams =
  | '/Time CLT'
  | '/TIME ALFA'
  | '/TIME FÊNIX'
  | '/TIME THÉMIS'
  | '/TIME TITÃ'
  | '/TIME ZION'
  | '/Administradores'
  | '/Admin';
export type PermissionAction =
  | 'clt:view'
  | 'clt:create'
  | 'clt:edit'
  | 'clt:delete'
  | 'clt:detail'
  | 'clt:manage'
  | 'Servants:view'
  | 'Servants:create'
  | 'Servants:edit'
  | 'Servants:delete'
  | 'Servants:detail'
  | 'Servants:manage';